added folder change to examples and build our local storage mechanism for handling tickets.
