+++ examples/sites/devcon-issuer/react-example-negociator-authenticator/README.md
