+++ examples/sites/devcon-issuer/react-example-negociator/package.json
