+++ examples/sites/devcon-issuer/react-example-negociator/public/favicon.ico
