+++ examples/sites/devcon-issuer/react-example-negociator/public/index.html
