+++ examples/sites/devcon-issuer/react-example-negociator/public/logo192.png
