+++ examples/sites/devcon-issuer/react-example-negociator/public/logo512.png
