+++ examples/sites/devcon-issuer/react-example-negociator/public/manifest.json
