+++ examples/sites/devcon-issuer/react-example-negociator/public/mock.png
