+++ examples/sites/devcon-issuer/react-example-negociator/src/App.test.js
