+++ examples/sites/devcon-issuer/react-example-negociator/src/Card/index.js
@@ -2,10 +2,8 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
-import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
-import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
 import './Card.css';
 
@@ -18,7 +16,7 @@ const useStyles = makeStyles({
   },
 });
 
-function MediaCard({ tokenInstance, selectVIPEventHandler }) {
+function MediaCard({ tokenInstance }) {
   const classes = useStyles();
   return (
     <Card className="card">
@@ -30,23 +28,16 @@ function MediaCard({ tokenInstance, selectVIPEventHandler }) {
         />
         <CardContent>
           <Typography gutterBottom variant="h5" component="h2">
-            {tokenInstance.token.ticketClass}
+            {tokenInstance.ticketClass}
           </Typography>
           <Typography variant="body2" color="textSecondary" component="p">
-            Ticket Id: {tokenInstance.token.ticketId}
+            Ticket Id: {tokenInstance.ticketId}
           </Typography>
           <Typography variant="body2" color="textSecondary" component="p">
-            Conference Id: {tokenInstance.token.conferenceId}
+            Devcon Id: {tokenInstance.devconId}
           </Typography>
         </CardContent>
       </CardActionArea>
-      <CardActions>
-        {tokenInstance.token.ticketClass === "VIP" &&
-          <Button size="small" color="primary" onClick={e => selectVIPEventHandler({ ticket: tokenInstance })}>
-            ENTER VIP
-        </Button>
-        }
-      </CardActions>
     </Card>
   );
 }
