+++ examples/sites/devcon-issuer/react-example-negociator/src/Modal/index.js
