+++ examples/sites/devcon-issuer/react-example-negociator/src/TokenScript/Authenticator.js
