+++ examples/sites/devcon-issuer/react-example-negociator/src/TokenScript/tokenMock.js
