+++ examples/sites/devcon-issuer/react-example-negociator/src/index.js
