+++ examples/sites/devcon-issuer/react-example-negociator/src/reportWebVitals.js
