+++ examples/sites/devcon-issuer/react-example-negociator/src/setupTests.js
