--- examples/sites/devcon-issuer/react-example/src/TokenScript/contractMock.js
@@ -1,5 +1,0 @@
-export const ContractMock = {
-  challenge: function async() {
-    return true
-  }
-}
